rect and dot lists

+++ pynsn/_lib/typing.py
@@ -4,4 +4,4 @@ from typing import List, Sequence, Union
 from numpy import int_
 from numpy.typing import NDArray
 
-IntOVector = Union[int, List[int], Sequence[int_], NDArray[int_]]
+IntOVector = Union[int, List[int], Sequence[int], NDArray[int_]]
